backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
django-filter==24.3
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
